[#597] Changes following rebase.
