+++ core/src/main/java/com/datastax/oss/driver/internal/core/channel/ProtocolInitHandler.java
@@ -40,8 +40,6 @@ import com.datastax.oss.driver.internal.core.protocol.BytesToSegmentDecoder;
 import com.datastax.oss.driver.internal.core.protocol.FrameDecoder;
 import com.datastax.oss.driver.internal.core.protocol.FrameEncoder;
 import com.datastax.oss.driver.internal.core.protocol.FrameToSegmentEncoder;
-import com.datastax.oss.driver.internal.core.protocol.ProtocolFeatureManager;
-import com.datastax.oss.driver.internal.core.protocol.ProtocolFeatureParser;
 import com.datastax.oss.driver.internal.core.protocol.ProtocolFeatureStore;
 import com.datastax.oss.driver.internal.core.protocol.SegmentToBytesEncoder;
 import com.datastax.oss.driver.internal.core.protocol.SegmentToFrameDecoder;
@@ -71,7 +69,6 @@ import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 import java.util.Objects;
-import java.util.Optional;
 import net.jcip.annotations.NotThreadSafe;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -97,7 +94,6 @@ class ProtocolInitHandler extends ConnectInitHandler {
   private String logPrefix;
   private ChannelHandlerContext ctx;
   private final boolean querySupportedOptions;
-  private ProtocolFeatureManager protocolFeatureManager;
   private ProtocolFeatureStore featureStore;
 
   /**
@@ -196,12 +192,6 @@ class ProtocolInitHandler extends ConnectInitHandler {
           if (featureStore != null) {
             featureStore.populateStartupOptions(startupOptions);
           }
-          Optional.ofNullable(protocolFeatureManager)
-              .ifPresent(m -> m.optionallyAddLwtInfoOption(startupOptions));
-          Optional.ofNullable(protocolFeatureManager)
-              .ifPresent(m -> m.optionallyAddTabletInfoOption(startupOptions));
-          Optional.ofNullable(protocolFeatureManager)
-              .ifPresent(m -> m.optionallyAddMetadataIdOption(startupOptions));
           return request = new Startup(startupOptions);
         case GET_CLUSTER_NAME:
           return request = CLUSTER_NAME_QUERY;
@@ -231,18 +221,11 @@ class ProtocolInitHandler extends ConnectInitHandler {
           ProtocolUtils.opcodeString(response.opcode));
       try {
         if (step == Step.OPTIONS && response instanceof Supported) {
-          channel.attr(DriverChannel.OPTIONS_KEY).set(((Supported) response).options);
-          Supported res = (Supported) response;
-          featureStore = ProtocolFeatureStore.parseSupportedOptions(res.options);
-          featureStore.storeInChannel(channel);
           Supported supported = (Supported) response;
-          ProtocolFeatureParser featureParser =
-              ProtocolFeatureParser.Builder.fromOptions(supported).build();
-          protocolFeatureManager = featureParser.parse();
-          protocolFeatureManager.updateOptionsAttributeForChannel(channel);
-          protocolFeatureManager.updateShardingInfoAttributeForChannel(channel);
-          protocolFeatureManager.updateLwtInfoAttributeForChannel(channel);
-          maybeUpdatePipelineWithProtocolOptions(protocolFeatureManager.isMetadataIdEnabled());
+          channel.attr(DriverChannel.OPTIONS_KEY).set(supported.options);
+          featureStore = ProtocolFeatureStore.parseSupportedOptions(supported.options);
+          featureStore.storeInChannel(channel);
+          maybeUpdatePipelineWithProtocolOptions(featureStore.isMetadataIdEnabled());
           step = Step.STARTUP;
           send();
         } else if (step == Step.STARTUP && response instanceof Ready) {
@@ -457,21 +440,21 @@ class ProtocolInitHandler extends ConnectInitHandler {
       ProtocolFeatures protocolFeatures = new ProtocolFeatures();
       protocolFeatures.addFeature(ProtocolFeatures.Feature.SCYLLA_USE_METADATA_ID);
       int maxFrameLength =
-        (int)
-          context
-            .getConfig()
-            .getDefaultProfile()
-            .getBytes(DefaultDriverOption.PROTOCOL_MAX_FRAME_LENGTH);
+          (int)
+              context
+                  .getConfig()
+                  .getDefaultProfile()
+                  .getBytes(DefaultDriverOption.PROTOCOL_MAX_FRAME_LENGTH);
 
       ChannelPipeline pipeline = ctx.pipeline();
       pipeline.replace(
-        ChannelFactory.FRAME_TO_BYTES_ENCODER_NAME,
-        ChannelFactory.FRAME_TO_BYTES_ENCODER_NAME,
-        new FrameEncoder(context.getFrameCodec(), protocolFeatures, maxFrameLength));
+          ChannelFactory.FRAME_TO_BYTES_ENCODER_NAME,
+          ChannelFactory.FRAME_TO_BYTES_ENCODER_NAME,
+          new FrameEncoder(context.getFrameCodec(), protocolFeatures, maxFrameLength));
       pipeline.replace(
-        ChannelFactory.BYTES_TO_FRAME_DECODER_NAME,
-        ChannelFactory.BYTES_TO_FRAME_DECODER_NAME,
-        new FrameDecoder(context.getFrameCodec(), protocolFeatures, maxFrameLength));
+          ChannelFactory.BYTES_TO_FRAME_DECODER_NAME,
+          ChannelFactory.BYTES_TO_FRAME_DECODER_NAME,
+          new FrameDecoder(context.getFrameCodec(), protocolFeatures, maxFrameLength));
     }
   }
